+++ src/features/game/Game.tsx
@@ -6,7 +6,6 @@ import { Hud } from "features/hud/Hud";
 import { Crops } from "features/crops/Crops";
 import { Blacksmith } from "features/blacksmith/Blacksmith";
 import { Mail } from "features/mail/Mail";
-import { MailProvider } from "features/mail/MailProvider";
 import { Water } from "features/water/Water";
 
 import { Context } from "./GameProvider";
@@ -33,9 +32,7 @@ export const Game: React.FC = () => {
       <Hud />
 
       <Blacksmith />
-      <MailProvider>
-        <Mail />
-      </MailProvider>
+      <Mail />
       <Crops />
       <Water />
     </>
