+++ src/features/mail/types/message.ts
@@ -1,0 +1,6 @@
+export type Message = {
+  id: string;
+  title?: string;
+  body: string;
+  unread: boolean;
+};
